refactor: handleRequest --> hreq
